requirements.txt
pygame
python-dotenv
nltk